PERF: Revert to shallow reactive
Since witching from asyncComputed to ref (shallow to deep reactive)
added performance issues, this commit introduces the new compositionApi
shallowRef and instead of setting the installedState property, perform a
full reload of the extension details from manager.

This means the button change will only work when loaded inside slicer
and actually installed. This means that directly mocking the behavior
calling the function "app.setExtensionButtonState" will not update the
state of the button.